feat(disabled): add support for disabled menu option #25
This update supports disabling the click event, enter
and spacebar key from selecting the menu option.

It is known at this point that the user can tab to
the disabled menu item. This will be addressed
with a refactor of how tab order is managed in a menu.

Changes to be committed:
modified:   src/auro-menu-option.scss
modified:   src/auro-menu.js
modified:   src/style.scss