feat(multi-select): add support for selection option #199
The core of this commit is updating the ability of the element to know
the difference between the default single-select and the multi-select
option. With the multi-select options selected, the value of the menu
selections will be kept in an array versus a single string.

Changes to be committed:
modified:   src/auro-menu.js
modified:   src/style-base.scss
modified:   src/style-menuoption.scss